data_ingestion/requirements.txt
requests
kafka-python
python-dotenv
riotwatcher
